refactor(extension): extract cover-on-save to dedicated module
