+++ vscode-gotest/src/coverage.ts
@@ -1,23 +1,11 @@
 import * as vscode from "vscode";
 import * as path from "node:path";
-import { rm, mkdtemp, readFile } from "node:fs/promises";
-import { tmpdir } from "node:os";
 import type { GoTestController } from "./testController.js";
 import type { DiscoveryCache } from "./discovery.js";
 import type { CoverageStore } from "./coverageStore.js";
-import {
-  buildCliCommand,
-  formatCliCommand,
-  scopedConfig,
-} from "./cli.js";
-import {
-  collectItems,
-  groupByPackage,
-  spawnTestProcess,
-  buildRunFilter,
-} from "./runnerUtils.js";
+import { scopedConfig } from "./cli.js";
+import { collectItems, groupByPackage, buildRunFilter } from "./runnerUtils.js";
 import { executeBatch } from "./batchRunner.js";
-import { runGoToolCoverFunc } from "./coverageUtils.js";
 
 export interface ParsedFileCoverage {
   absPath: string;
@@ -246,9 +234,6 @@ export function parseFuncCoverage(
 
 export class CoverageRunner implements vscode.Disposable {
   private activeRun: vscode.CancellationTokenSource | undefined;
-  private activePackageRun: vscode.CancellationTokenSource | undefined;
-  private packageRunQueue = Promise.resolve();
-  private pendingPackages = new Set<string>();
 
   constructor(
     private readonly controller: GoTestController,
@@ -459,97 +444,8 @@ export class CoverageRunner implements vscode.Disposable {
     return result.stdout;
   }
 
-  async runPackage(importPath: string): Promise<void> {
-    if (this.pendingPackages.has(importPath)) return;
-    this.pendingPackages.add(importPath);
-    this.packageRunQueue = this.packageRunQueue.then(async () => {
-      this.pendingPackages.delete(importPath);
-      await this.executePackageRun(importPath);
-    });
-    return this.packageRunQueue;
-  }
-
-  private async executePackageRun(importPath: string): Promise<void> {
-    this.activePackageRun?.cancel();
-    const cts = new vscode.CancellationTokenSource();
-    this.activePackageRun = cts;
-
-    const pkg = this.cache.getPackage(importPath);
-    if (!pkg) return;
-    const workspaceDir = this.cache.getWorkspaceDir(importPath);
-    if (!workspaceDir) return;
-
-    const config = scopedConfig(workspaceDir);
-    const testFlags = config.get<string[]>("testFlags") ?? [];
-    let coverFile: string | undefined;
-
-    try {
-      const tmpDir = await mkdtemp(path.join(tmpdir(), "gotest-cov-"));
-      coverFile = path.join(tmpDir, "cover.out");
-
-      const cliArgs: string[] = [
-        "-count=1",
-        "-covermode=atomic",
-        `-coverprofile=${coverFile}`,
-        importPath,
-      ];
-      cliArgs.push(...testFlags);
-
-      const cmd = await buildCliCommand(
-        cliArgs,
-        workspaceDir,
-        this.outputChannel,
-      );
-      this.outputChannel.appendLine(`[coverage:save] ${formatCliCommand(cmd)}`);
-
-      await spawnTestProcess(
-        cmd.bin,
-        cmd.args,
-        workspaceDir,
-        cts.token,
-        this.outputChannel,
-        "coverage",
-      );
-
-      if (cts.token.isCancellationRequested) return;
-
-      const coverContent = await readFile(coverFile, "utf-8");
-      let funcOutput: string | undefined;
-      try {
-        funcOutput = await runGoToolCoverFunc(coverFile, workspaceDir);
-      } catch {
-        this.outputChannel.appendLine(
-          "[coverage:save] go tool cover -func failed",
-        );
-      }
-      this.store.update(importPath, coverContent, funcOutput);
-    } catch (err) {
-      const message = err instanceof Error ? err.message : String(err);
-      this.outputChannel.appendLine(`[coverage:save] failed: ${message}`);
-      return;
-    } finally {
-      if (this.activePackageRun === cts) this.activePackageRun = undefined;
-      cts.dispose();
-      if (coverFile)
-        rm(path.dirname(coverFile), { recursive: true, force: true }).catch(
-          () => {},
-        );
-    }
-
-    const request = new vscode.TestRunRequest();
-    const run = this.controller.createTestRun(request, "Cover on Save");
-    const { coverages: allCoverages } = this.store.buildFileCoverages(this.cache);
-    for (const fc of allCoverages) {
-      run.addCoverage(fc);
-    }
-    run.end();
-    await this.store.save();
-  }
-
   dispose(): void {
     this.activeRun?.cancel();
     this.activeRun = undefined;
-    this.activePackageRun?.cancel();
-    this.activePackageRun = undefined;
   }
 }
